+++ apps/vision-demo/CHANGELOG.md
@@ -1,5 +1,12 @@
 # rdk-vision-demo
 
+## 0.0.3
+
+### Patch Changes
+
+- Updated dependencies [[`8b870a5`](https://github.com/omnidotdev/rdk/commit/8b870a5ce26d32d141299083bf6274111ed6f1ab)]:
+  - @omnidotdev/rdk@0.12.1
+
 ## 0.0.2
 
 ### Patch Changes